Extract testing library from klb tests
