+++ tests/azure/availset_test.go
@@ -5,9 +5,9 @@ import (
 	"math/rand"
 	"testing"
 
-	"github.com/NeowayLabs/klb/tests/azure"
-	"github.com/NeowayLabs/klb/tests/azure/fixture"
-	"github.com/NeowayLabs/klb/tests/nash"
+	"github.com/NeowayLabs/klb/tests/lib/azure"
+	"github.com/NeowayLabs/klb/tests/lib/azure/fixture"
+	"github.com/NeowayLabs/klb/tests/lib/nash"
 
 	"github.com/NeowayLabs/nash/sh"
 )
