+++ tests/azure/doc.go
@@ -1,0 +1,3 @@
+//Package azure contains tests for Azure
+//These are tests for klb nash code
+package azure
