+++ tests/lib/azure/availset.go
