+++ tests/lib/azure/doc.go
@@ -1,0 +1,3 @@
+//Package azure provides useful functions to write
+//tests that validate infrastructure built on Azure
+package azure
