+++ tests/lib/azure/resourcegroup.go
