+++ tests/lib/azure/session.go
