+++ tests/lib/nash/nash.go
